+++ changes.md
@@ -6,3 +6,10 @@
 - :warning: removed type `MyCartSetDirectDiscountsAction`
 </details>
 
+
+<details>
+<summary>Added Property(s)</summary>
+
+- added property `perMethodTaxRate` to type `CustomLineItem`
+</details>
+